cleaning up some things and organising a bit
