+++ Makefile
@@ -6,40 +6,16 @@
 #    By: tmullan <tmullan@student.codam.nl>           +#+                      #
 #                                                    +#+                       #
 #    Created: 2021/02/15 13:01:13 by tmullan       #+#    #+#                  #
-#    Updated: 2021/04/20 12:26:31 by zenotan       ########   odam.nl          #
+#    Updated: 2021/04/21 22:47:46 by zenotan       ########   odam.nl          #
 #                                                                              #
 # **************************************************************************** #
 
 NAME = ghostshell
 
-GNL = gnl/get_next_line.c \
-		gnl/get_next_line_utils.c
-
-# LFT = ft_putstr_fd.c \
-# 		ft_putnbr_fd.c \
-# 		ft_putchar_fd.c \
-# 		ft_strlen.c \
-# 		ft_split.c \
-# 		ft_memcpy.c \
-# 		ft_strchr.c \
-# 		ft_strdup.c \
-# 		ft_strlcpy.c \
-# 		ft_substr.c \
-# 		ft_strjoin.c \
-# 		ft_strnstr.c \
-# 		ft_strcmp.c \
-# 		ft_bzero.c \
-# 		ft_lstadd_back_bonus.c \
-# 		ft_lstclear_bonus.c \
-# 		ft_lstdelone_bonus.c \
-# 		ft_lstiter_bonus.c \
-# 		ft_lstnew_bonus.c \
-# 		ft_lstsize_bonus.c \
-# 		ft_lstadd_front_bonus.c
-
 ERR = error/error.c
 
 UTIL = struct_utils.c \
+		dlist.c \
 		lst_utils.c \
 		history_utils.c \
 		env_utils.c \
@@ -47,10 +23,11 @@ UTIL = struct_utils.c \
 		free_all.c \
 		ft_strjoinfree.c
 
-LEX = read_input.c \
+PAR = parser.c \
+		parser_utils.c \
+		handle_env.c \
+		read_input.c \
 		lexer.c
-
-PAR = parser.c
 
 SHEL = shell_exec.c \
 		prog_launch.c \
@@ -60,10 +37,6 @@ SHEL = shell_exec.c \
 TEST = test.c
 
 PAR_PREFIX = $(addprefix parser/, $(PAR))
-
-# LFT_PREFIX = $(addprefix lft_utils/, $(LFT))
-
-LEX_PREFIX = $(addprefix lexer/, $(LEX))
 
 UTIL_PREFIX = $(addprefix utils/, $(UTIL))
