+++ includes/ghostshell.h
@@ -6,7 +6,7 @@
 /*   By: tmullan <tmullan@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/02/15 13:04:04 by tmullan       #+#    #+#                 */
-/*   Updated: 2021/04/20 12:35:35 by zenotan       ########   odam.nl         */
+/*   Updated: 2021/04/21 22:50:31 by zenotan       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,6 @@
 # include <sys/types.h>
 # include <sys/stat.h>
 # include <sys/wait.h>
-# include "get_next_line.h"
 # include "libft.h"
 # include "reins.h"
 
@@ -64,12 +63,6 @@ enum	e_return
 	SYNTAX_ERR = 258
 };
 
-// typedef struct		s_list
-// {
-// 	void			*content;
-// 	struct s_list	*next;
-// }					t_list;
-
 typedef struct		s_dlist
 {
 	void			*content;
@@ -100,9 +93,6 @@ typedef struct		s_shell
 	t_list	*commands;
 	t_dlist	*tokens;
 	t_reins	*reins;
-	// char	**tokens;
-	// comands
-	// status
 	pid_t	pid;
 	char	**env;
 	char	*line;
@@ -113,6 +103,7 @@ typedef struct		s_shell
 	int		error;
 }					t_shell;
 
+//---------------------------------shell_exec---------------------------------//
 // built-in functions
 int		run_echo(t_cmd *cmd, t_shell **ghost);
 int		run_cd(t_cmd *cmd, t_shell **ghost);
@@ -141,58 +132,35 @@ int		redir_muti(void *file_struct);
 int		pipe_exec(t_list *command, t_shell **ghost);
 int		first_cmd(pid_t pid, t_list *command, t_shell **ghost, int fd_in);
 
-// // lft_utils
-// size_t	ft_strlen(const char *s);
-// void	*ft_memcpy(void *dst, const void *src, size_t n);
-// void	ft_putstr_fd(char *str, int fd);
-// int		ft_strcmp(const char *str1, const char *str2);
-// void	ft_putnbr_fd(int n, int fd);
-// void	ft_putchar_fd(char c, int fd);
-// char	**ft_split(char const *s, char c);
-// char	*ft_strchr(const char *s, int c);
-// char	*ft_strdup(const char *s1);
-// size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
-// char	*ft_substr(char const *s, unsigned int start, size_t len);
-// char	*ft_strjoin(char *s1, char const *s2);
-// int		ft_strncmp(const char *s1, const char *s2, size_t n);
-// char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
-// void	ft_bzero(void *s, size_t n);
-// int		ft_isdigit(int c);
-// int		ft_isalpha(int c);
-// int		ft_isalnum(int c);
-// int		ft_atoi(const char *str);
-
-//list
-void	ft_lstadd_back(t_list **alst, t_list *new);
-void	ft_lstadd_front(t_list **alst, t_list *new);
-void	ft_lstclear(t_list **lst, void (*del)(void*));
-void	ft_lstdelone(t_list *lst, void (*del)(void*));
-void	ft_lstiter(t_list *lst, void (*f)(void *));
-t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-t_list	*ft_lstlast(t_list *lst);
-t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-t_list	*ft_lstnew(void *content);
-int		ft_lstsize(t_list *lst);
-char	**list_to_arr(t_list *tokens);
-
-// error
+//-----------------------------------error------------------------------------//
+//error.c
 void	error_handler(t_shell **ghost, int error_code, char *error_message, char *arg);
 void	cmd_notfound(t_cmd *cmd, int flag, t_shell **ghost);
 
-// lexer
+//-----------------------------------parser-----------------------------------//
+//parser.c
+void	parser_loop(t_shell **ghost);
+char 	*remove_quotes(t_shell **ghost, char *str, int len);
+
+//lexer.c
 void	read_line(t_shell **ghost);
 int		up_function(t_input *line, char *buf, t_hook *hook);
 int		down_function(t_input *line, char *buf, t_hook *hook);
-// void	lexer(t_shell **shell, char *input);
 void	lexer(t_shell **shell);
 
-// parser
-void	parser(t_shell **ghost);
+//handle_env.c
 char	**get_envp(char **envp);
-int		check_meta(char *str);
 void	handle_env(t_shell **ghost, char **content);
 
-// struct_utils
+//parser_utils.c
+int		check_meta(char *str);
+int		check_quote(t_shell **ghost);
+int		check_colon(t_shell **ghost, t_cmd **command);
+int		check_seperator(t_shell **ghost, t_cmd **command);
+int		check_redir(t_shell **ghost, t_cmd **command);
+
+//------------------------------------utils-----------------------------------//
+//struct_utils.c
 t_shell	*init_shell(char **env);
 t_redir	*new_redir(t_shell **ghost, char *file, int type);
 t_cmd	*new_command();
@@ -201,9 +169,23 @@ void	del_commands(void *list);
 void	del_content(void *content);
 void	del_darray(char **str);
 
-// lst_utils
+//lst_utils.c
 void	*copy_data(void	*data);
-// char	**dlist_to_arr(t_dlist *tokens);
+char	**list_to_arr(t_list *tokens);
+
+//history_utils.c
+void	store_command(t_shell **ghost, char *line);
+void	init_reins(t_shell **ghost);
+void	pass_param(void *param);
+void	edit_content(t_dlist **node, char *line, int size);
+
+//env_utils.c
+int 	valid_val(char *str);
+int		valid_word(char *str);
+t_dlist	*split_env(char *str);
+int		replace_env_quoted(t_shell **ghost, char **input, int i);
+
+//dlist.c
 t_dlist	*ft_dlstnew(void *content);
 void	ft_dlstadd_front(t_dlist **alst, t_dlist *new);
 void	ft_dlstclear(t_dlist **lst, void (*del)(void *));
@@ -213,30 +195,17 @@ void	ft_dlstdelone(t_dlist **lst, int position, void (*del)(void *));
 void	dreplace(t_dlist **lst, t_dlist *insert, int pos, void (*del)(void *));
 void	ft_dlsreversetiter(t_dlist *lst, void (*f)(void *));
 int		ft_lstredir(t_list *lst, int (*f)(void *));
-
 void	ft_dlstadd_back(t_dlist **alst, t_dlist *new);
 void	ft_dlstiter(t_dlist *lst, void (*f)(void *));
 void	del_ghost(t_shell **ghost);
 t_dlist	*ft_dlstfirst(t_dlist *lst);
 
-//	history_utils
-void	store_command(t_shell **ghost, char *line);
-void	init_reins(t_shell **ghost);
-void	pass_param(void *param);
-void	edit_content(t_dlist **node, char *line, int size);
-
-//env_utils
-int 	valid_val(char *str);
-int		valid_word(char *str);
-t_dlist	*split_env(char *str);
-int		replace_env_quoted(t_shell **ghost, char **input, int i);
-
-// Utils
-char	**arr_addback(char **arr, char *str);
+//tur_utils.c
 void	free_all(t_shell **ghost);
+char	**arr_addback(char **arr, char *str);
 char	*ft_strjoinfree(char *s1, char const *s2);
 
-//debug
+//-----------------------------------debug-----------------------------------//
 void	print_data(void *data);
 void	print_cmd(t_cmd *data);
 void	ft_cmd_lstiter(t_list *lst, void (*f)(t_cmd *));
