+++ test.c
@@ -6,7 +6,7 @@
 /*   By: ztan <ztan@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/04/05 17:12:59 by ztan          #+#    #+#                 */
-/*   Updated: 2021/04/15 21:32:21 by zenotan       ########   odam.nl         */
+/*   Updated: 2021/04/21 22:26:03 by zenotan       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -229,3 +229,77 @@ void	test()
 	env_test(&list);
 
 }
+
+// char	*get_env(t_shell **ghost, char *name)
+// {
+// 	char **list;
+// 	int i;
+// 	int j;
+
+// 	if ((*ghost)->env)
+// 		list = (*ghost)->env;
+// 	i = 0;
+// 	while (list[i])
+// 	{
+// 		j = 0;
+// 		while (name[j] == list[i][j])
+// 			j++;
+// 		if (list[i][j] == '=')
+// 			return (ft_substr(list[i], j + 1, ft_strlen(list[i])));
+// 		i++;
+// 	}
+// 	return (NULL);
+// }
+
+// void	handle_env(t_shell **ghost, char **content)
+// {
+// 	free(*content);
+// 	*content = get_env(ghost, *content);
+// 	if (!*content)
+// 		error_handler(ghost, INVALID_ENV,  "Failed to recognize env", *content);
+// }
+
+// // lft_utils
+// size_t	ft_strlen(const char *s);
+// void	*ft_memcpy(void *dst, const void *src, size_t n);
+// void	ft_putstr_fd(char *str, int fd);
+// int		ft_strcmp(const char *str1, const char *str2);
+// void	ft_putnbr_fd(int n, int fd);
+// void	ft_putchar_fd(char c, int fd);
+// char	**ft_split(char const *s, char c);
+// char	*ft_strchr(const char *s, int c);
+// char	*ft_strdup(const char *s1);
+// size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize);
+// char	*ft_substr(char const *s, unsigned int start, size_t len);
+// char	*ft_strjoin(char *s1, char const *s2);
+// int		ft_strncmp(const char *s1, const char *s2, size_t n);
+// char	*ft_strnstr(const char *haystack, const char *needle, size_t len);
+// void	ft_bzero(void *s, size_t n);
+// int		ft_isdigit(int c);
+// int		ft_isalpha(int c);
+// int		ft_isalnum(int c);
+// int		ft_atoi(const char *str);
+
+// # LFT = ft_putstr_fd.c \
+// # 		ft_putnbr_fd.c \
+// # 		ft_putchar_fd.c \
+// # 		ft_strlen.c \
+// # 		ft_split.c \
+// # 		ft_memcpy.c \
+// # 		ft_strchr.c \
+// # 		ft_strdup.c \
+// # 		ft_strlcpy.c \
+// # 		ft_substr.c \
+// # 		ft_strjoin.c \
+// # 		ft_strnstr.c \
+// # 		ft_strcmp.c \
+// # 		ft_bzero.c \
+// # 		ft_lstadd_back_bonus.c \
+// # 		ft_lstclear_bonus.c \
+// # 		ft_lstdelone_bonus.c \
+// # 		ft_lstiter_bonus.c \
+// # 		ft_lstnew_bonus.c \
+// # 		ft_lstsize_bonus.c \
+// # 		ft_lstadd_front_bonus.c
+
+// # LFT_PREFIX = $(addprefix lft_utils/, $(LFT))
